How was this working?

+++ phoenix-2.0/tests/matrix_multiply/matrix_multiply-pthread.c
@@ -122,7 +122,7 @@ void matrixmult_splitter(void *data_in)
     /* Barrier, wait for all threads to finish */
     for (i = 0; i < num_procs; i++)
     {
-        int ret_val;
+        long ret_val;
         CHECK_ERROR(pthread_join(tid[i], (void **)(void*)&ret_val) != 0);
 	  CHECK_ERROR(ret_val != 0);
     }
